New verify func to make secure when using public keys
If using asymmetric keys, and the public key is actually publicly
available, then you don't want the user supplying the token to be
able to determine the algorithm used to verify the signature. They
could use the public key to create a fake signature using an HMAC
algorithm, and set that as the 'alg' value in the header.

To avoid that, use the new 'verify' function, where you supply the
algorithm to use. It checks that the token's header matches that,
rather than just taking the token's word for it.

Use like this for example:
    local body = assert(jwt.verify(token, "RS256", pubkey))

If you're using an HMAC algorithm anyway, so your key is kept secret,
then there is no problem continuing to use the 'decode' function.

The issue is explained in this blog post:
https://auth0.com/blog/critical-vulnerabilities-in-json-web-token-libraries/